a few 64-bit compatibility changes for vs2005
git-svn-id: file://localhost/home/tolik/svnrepo/trunk@585 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ opencv/src/highgui/bitstrm.cpp
@@ -600,7 +600,7 @@ bool bsCreateDecodeHuffmanTable( const int* src, short* table, int max_size )
     int  i, k;
     
     /* calc bit depths of sub tables */
-    memset( sub_tables, 0, (1 << first_bits)*sizeof(sub_tables[0]) );
+    memset( sub_tables, 0, ((size_t)1 << first_bits)*sizeof(sub_tables[0]) );
     for( i = 1, k = 1; src[k] >= 0; i++ )
     {
         int code_count = src[k++];
